+++ ChangeLog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- FIX : Lors de la mise en pause avec tasklist, la progression déclarée calculée automatiquement est maintenant arrondie au multiple de 5 supérieur le plus proche,
+        car sinon en cas de valeur comme 4 ou 17 % d'avancement, le pourcentage d'avancement actuel n'est pas présélectionné sur la vue fullcalendar des tâches *23/04/2021* eprolor_2.1
 - NEW : Ajout du client sur la vue tasklist sur la ligne principale de la tâche en premier (avant le libellé, etc.) *23/04/2021* - eprolor_2.1
 - NEW : Suite point avec Thibault, lorsque le module ordo est désactivé, on ne veut voir dans tasklist que les tâches dont les date de début est aujourd'hui *06/04/2021* - eprolor_2.1
 - NEW : ajout date prevue livraison la plus proche sur vue planning et tasklist (T3027) *07/04/2021* - eprolor_2.1